src/requirements.txt
GitPython 
radon
requests

